+++ Kernel/Arch/aarch64/TrapFrame.h
@@ -8,6 +8,7 @@
 #pragma once
 
 #include <AK/Platform.h>
+#include <AK/StdLibExtras.h>
 #include <AK/Types.h>
 #include <Kernel/Arch/RegisterState.h>
