+++ Kernel/Arch/riscv64/RegisterState.h
@@ -12,6 +12,8 @@
 #include <Kernel/Security/ExecutionMode.h>
 
 #include <AK/Platform.h>
+#include <AK/StdLibExtras.h>
+
 VALIDATE_IS_RISCV64()
 
 namespace Kernel {
