+++ Kernel/Arch/riscv64/TrapFrame.h
@@ -9,6 +9,8 @@
 #include <Kernel/Arch/RegisterState.h>
 
 #include <AK/Platform.h>
+#include <AK/StdLibExtras.h>
+
 VALIDATE_IS_RISCV64()
 
 namespace Kernel {
