+++ Kernel/Devices/Storage/SD/Commands.h
@@ -6,6 +6,7 @@
 
 #pragma once
 
+#include <AK/StdLibExtras.h>
 #include <AK/Types.h>
 
 namespace Kernel::SD {
