+++ Userland/Libraries/LibC/arch/aarch64/fenv.cpp
@@ -4,6 +4,7 @@
  * SPDX-License-Identifier: BSD-2-Clause
  */
 
+#include <AK/Assertions.h>
 #include <AK/Types.h>
 #include <fenv.h>
